make the interpolation work (redimensioning the wave was plain wrong)

+++ VC8/brickletconverter.h
@@ -2,4 +2,5 @@
 
 #include "header.h"
 
-int createWaves(DataFolderHandle dataFolderHandle, const char *baseName,int brickletID,double resampleSize, std::string &fullPathOfCreatedWaves);
+int createWaves(DataFolderHandle dataFolderHandle, const char *baseName,int brickletID,int resampleSize, std::string &fullPathOfCreatedWaves);
+
